requirements.txt
Django==1.7.1
ipython
python-social-auth
psycopg2
djangorestframework
pillow
djangorestframework-gis
uwsgi
django-suit
django-cors-headers
django-oauth-toolkit
